Get rid of proc_common.cpp file
It was being included in 4 places.
All proc_common.cpp content is now in proc.cpp

Reorganization will follow.
